Split native DTO projection tests
